feat(maintainer): Update keyspace handling and mocks
- Replaced `common.DefaultKeyspaceID` with `common.DefaultKeyspace` in `NewController` calls.
- Updated mock schema store to use `eventservice.NewMockSchemaStore()` and `SetTables()` for better testability.
- Modified `GetTableInfo`, `GetAllPhysicalTables`, `GetTableDDLEventState`, `RegisterTable`, `UnregisterTable`, and `FetchTableDDLEvents` in `mockSchemaStore` to accept `common.KeyspaceMeta` instead of `uint32` keyspace ID.
- Updated `RegisterKeyspace` in `mockSchemaStore` to accept `common.KeyspaceMeta`.
- Adjusted tests in `maintainer_controller_test.go` and `maintainer_manager_test.go` to reflect these changes.
- Removed `mockSchemaStore` struct from `maintainer_manager_test.go` and `maintainer_test.go` as it's now defined in `pkg/eventservice/test_helper.go`.
- Ensured that `common.DefaultKeyspaceID` is now a `uint32` to align with `common.KeyspaceMeta.ID`.
- Updated `TestMaintainerSchedulesNodeChanges` and `TestMaintainerBootstrapWithTablesReported` to correctly handle keyspace ID for next-gen environments.

+++ pkg/common/types.go
@@ -363,7 +363,7 @@ const (
 	// DefaultKeyspaceNamme is the default keyspace value,
 	// all the old changefeed will be put into default keyspace
 	DefaultKeyspaceNamme = "default"
-	DefaultKeyspaceID    = 0
+	DefaultKeyspaceID    = uint32(0)
 )
 
 var DefaultKeyspace = KeyspaceMeta{
